requirements.txt
# Основная библиотека для Telegram-бота
python-telegram-bot

# Для работы с базой данных SQLite (входит в стандартную библиотеку, но указываем зависимости)
# sqlite3 — встроен в Python, не требует установки

# Для HTTP-запросов (например, при интеграции с TTS)
requests==2.31.0

# Для обработки аудио (если будешь использовать TTS и конвертацию в MP3)
pydub==0.5.3

# Дополнительно: для логирования (встроен в Python, но можно уточнить)
# logging — встроен

# Дополнительно: если будешь использовать асинхронность (уже включено в python-telegram-bot)
# asyncio — встроен в Python

# Дополнительно: для генерации речи (опционально — Yandex, Google, ElevenLabs)
# Примеры (раскомментируй, если нужно):
# yandex-speechkit==0.1.4
# elevenlabs==0.3.2
# gtts==2.5.1